feat: supoort click block on doc mode
