+++ packages/presets/src/fragments/index.ts
@@ -3,7 +3,4 @@ export {
   FramePanel,
   registerFramePanelComponents,
 } from './frame-panel/frame-panel.js';
-export {
-  registerTOCPanelComponents,
-  TOCNotePanel,
-} from './toc-panel/toc-panel.js';
+export { registerTOCPanelComponents, TOCPanel } from './toc-panel/toc-panel.js';
